Changed Printing and Doc/View.

+++ Library/Library/ClipBoard.cpp
@@ -4,12 +4,12 @@
 
 #include "pch.h"
 #include "ClipBoard.h"
-#include "MessageBox.h"
 
 
 bool ClipBoard::load(TCchar* tc) {
 ToAnsi  ansi(tc);
-size_t  cbStr = ansi.length() + 1;
+//size_t  cbStr = ansi.length() + 1;
+int     cbStr = ansi.length() + 1;
 char*   p;
 HGLOBAL hData;
